requirements.txt
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
blis==0.4.1
bootstrap4==0.1.0
cached-property==1.5.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
cymem==2.0.3
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-adminlte2==0.4.3
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-phonenumber-field==4.0.0
django-profanity-filter==0.2.1
django-stdimage==5.1.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
joblib==0.16.0
more-itertools==8.4.0
murmurhash==1.0.2
numpy==1.19.1
oauthlib==3.1.0
ordered-set==3.1.1
ordered-set-stubs==0.1.3
phonenumbers==8.12.6
Pillow==7.2.0
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.2
progressbar2==3.51.4
psycopg2-binary==2.8.6
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.100
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
static3==0.7.0
thinc==7.4.1
threadpoolctl==2.1.0
tomlkit==0.5.11
tqdm==4.48.2
urllib3==1.25.10
wasabi==0.7.1
whitenoise==5.2.0
zipp==3.1.0
